fix some memory leak issues

+++ include/io.h
@@ -1,6 +1,6 @@
 #ifndef CSON_IO_H_
 #define CSON_IO_H_
 
-int read_file_content(const char* filename, char** output);
+unsigned long read_file_content(const char* filename, char** output);
 
 #endif // CSON_IO_H_
